refactor: move data format autoconfigurations to the geotools namespace
Refactor packages

* `org.geoserver.cloud.autoconfigure.rasterformats` -> `org.geotools.autoconfigure.rasterformats`
* `org.geoserver.cloud.autoconfigure.vectorformats` -> `org.geotools.autoconfigure.vectorformats`.
