+++ src/extensions/raster-formats/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.geoserver.cloud.autoconfigure.rasterformats.GridFormatFactoryFilteringAutoConfiguration,\
+org.geotools.autoconfigure.rasterformats.GridFormatFactoryFilteringAutoConfiguration,\
 org.geoserver.cloud.autoconfigure.cog.COGAutoConfiguration,\
 org.geoserver.cloud.autoconfigure.cog.COGWebUIAutoConfiguration
