+++ src/extensions/raster-formats/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,3 +1,3 @@
-org.geoserver.cloud.autoconfigure.rasterformats.GridFormatFactoryFilteringAutoConfiguration
+org.geotools.autoconfigure.rasterformats.GridFormatFactoryFilteringAutoConfiguration
 org.geoserver.cloud.autoconfigure.cog.COGAutoConfiguration
 org.geoserver.cloud.autoconfigure.cog.COGWebUIAutoConfiguration
