+++ src/extensions/vector-formats/src/main/java/org/geotools/autoconfigure/vectorformats/DataAccessFactoryFilterConfigProperties.java
@@ -3,7 +3,7 @@
  * application directory.
  */
 
-package org.geoserver.cloud.autoconfigure.vectorformats;
+package org.geotools.autoconfigure.vectorformats;
 
 import java.util.HashMap;
 import java.util.Map;
