+++ src/extensions/vector-formats/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-org.geoserver.cloud.autoconfigure.vectorformats.DataAccessFactoryFilteringAutoConfiguration
+org.geotools.autoconfigure.vectorformats.DataAccessFactoryFilteringAutoConfiguration
